requirements.txt
pymongo
feedparser
beautifulsoup4
aiohttp
python-dateutil
pytz
cloudscraper
